Cast PeerInfo port to uint16

+++ chia/timelord/timelord_launcher.py
@@ -12,6 +12,7 @@ from chia.types.peer_info import PeerInfo
 from chia.util.chia_logging import initialize_logging
 from chia.util.config import load_config
 from chia.util.default_root import DEFAULT_ROOT_PATH
+from chia.util.ints import uint16
 from chia.util.setproctitle import setproctitle
 
 active_processes: List = []
@@ -50,7 +51,7 @@ async def spawn_process(host: str, port: int, counter: int):
         try:
             dirname = path_to_vdf_client.parent
             basename = path_to_vdf_client.name
-            check_addr = PeerInfo(host, port)
+            check_addr = PeerInfo(host, uint16(port))
             if check_addr.is_valid():
                 resolved = host
             else:
